chore: @sharebook-layouts -> move DefaultLayout into a folder
